+++ COT/helpers/fatdisk.py
@@ -3,7 +3,7 @@
 # fatdisk.py - Helper for 'fatdisk'
 #
 # February 2015, Glenn F. Matthews
-# Copyright (c) 2013-2016 the COT project developers.
+# Copyright (c) 2013-2017 the COT project developers.
 # See the COPYRIGHT.txt file at the top-level directory of this distribution
 # and at https://github.com/glennmatthews/cot/blob/master/COPYRIGHT.txt.
 #
@@ -54,7 +54,7 @@ class FatDisk(Helper):
             helpers['port'].install_package('fatdisk')
             return
         elif platform.system() != 'Linux':
-            self.unsure_how_to_install()
+            raise self.unsure_how_to_install()
 
         # Fatdisk installation requires make
         helpers['make'].install()