+++ COT/ovf/ovf.py
@@ -2989,3 +2989,8 @@ class OVF(VMDescription, XML):
             (capacity, cap_units) = factor_bytes(capacity_bytes)
             disk.set(self.DISK_CAPACITY, capacity)
             disk.set(self.DISK_CAP_UNITS, cap_units)
+
+
+if __name__ == "__main__":
+    import doctest
+    doctest.testmod()